Start to work toward returning the server construct
And as such removing the closure for the server assignment. Once we have
a fixed names, we can return it and let the user wait for the server as
they wish. That will also make testing easier as we will have a way to
listen for the "server is now serving" event.

That being said, this commit does NOT compile as I still have to resolve
an issue with hyper complaining about a Result expected. Their trait
alias strategy make it really hard to understand what is going on :(